// src/data/models/signup.model.ts
export interface signupModel{
  title: string;
  name: string;
  password: string;
  day: string;
  month: string;
  year: string;
  newsletter: boolean;
  offers: boolean;
  firstName: string;
  lastName: string;
  company: string;
  address: string;
  address2: string;
  state: string;
  city: string;
  zipcode: string;
  mobileNumber: string;
}